fix: error atom added and images folder renamed to team
