+++ src/atoms/modals.ts
@@ -1,13 +1,13 @@
-import { ModalProps } from '@/components/ui/modal';
-import { atom } from 'recoil';
+// import { ModalProps } from '@/components/ui/modal';
+// import { atom } from 'recoil';
 
-export type ModalType = {
-    id: string;
-    type: 'logout' | 'confirm' | 'others';
-    props: ModalProps;
-};
+// export type ModalType = {
+//     id: string;
+//     type: 'logout' | 'confirm' | 'others';
+//     props: ModalProps;
+// };
 
-export const isModalOpenAtom = atom<Array<ModalType>>({
-    key: 'isModalOpenAtom',
-    default: [],
-});
+// export const isModalOpenAtom = atom<Array<ModalType>>({
+//     key: 'isModalOpenAtom',
+//     default: [],
+// });
