+++ src/components/Header/HeaderDesktop.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { Link, NavLink as RouterNavLink } from 'react-router-dom';
-import kodjLogo from '@/static/members/kodj_new.jpg';
+import kodjLogo from '@/static/team/kodj_new.jpg';
 import themeColors from '@/tools/themeColors';
 import Button from '@/components/Button/Button';
 import { HeaderProps } from '@/types';
