+++ src/static/team/bekhzod.png
