+++ src/static/team/javokhir.jpg
