+++ src/static/team/oybek.jpg
