+++ src/static/team/sardor.png
